Добавил обработку ошибки NotFoundException

+++ src/main/java/org/acme/service/impl/CarServiceImpl.java
@@ -2,6 +2,9 @@ package org.acme.service.impl;
 
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.inject.Inject;
+import jakarta.ws.rs.NotFoundException;
+import jakarta.ws.rs.WebApplicationException;
+import jakarta.ws.rs.core.Response;
 import org.acme.dto.CarDTO;
 import org.acme.entity.Car;
 import org.acme.mappers.CarMapper;
@@ -20,6 +23,9 @@ public class CarServiceImpl implements CarService {
     @Override
     public List<CarDTO> getAll() {
         List<Car> carList = carRepository.findAll().list();
+        if(carList.isEmpty()){
+            throw new NotFoundException();
+        }
         List<CarDTO> carDTOList = new ArrayList<>();
         for(Car car: carList){
             carDTOList.add(CarMapper.INSTANCE.toDTO(car));
@@ -29,8 +35,11 @@ public class CarServiceImpl implements CarService {
 
     @Override
     public CarDTO getById(Long id) {
-        Car car = carRepository.findById(id);
-        return CarMapper.INSTANCE.toDTO(car);
+        Car existingCar = carRepository.findById(id);
+        if(existingCar == null){
+            throw new NotFoundException();
+        }
+        return CarMapper.INSTANCE.toDTO(existingCar);
     }
 
     @Override
@@ -41,6 +50,10 @@ public class CarServiceImpl implements CarService {
 
     @Override
     public void delete(Long id) {
+        Car existingCar = carRepository.findById(id);
+        if(existingCar == null){
+            throw new NotFoundException();
+        }
         carRepository.deleteById(id);
     }
 
@@ -48,6 +61,9 @@ public class CarServiceImpl implements CarService {
     public void update(Long id, CarDTO carDTO) {
         Car carFromDTO = CarMapper.INSTANCE.toEntity(carDTO);
         Car existingCar = carRepository.findById(id);
+        if(existingCar == null){
+            throw new NotFoundException();
+        }
         existingCar.setName(carFromDTO.getName());
         existingCar.setColor(carFromDTO.getColor());
         existingCar.setMaxSpeed(carFromDTO.getMaxSpeed());
